Move assert_dirs_match() to test utils
